requirements.txt
absl-py==1.2.0
dm-haiku==0.0.7
ml_collections==0.1.1
numpy>=1.21.0
optax==0.1.3
jraph>=0.0.5.dev0
scikit-learn>=0.24.0
matplotlib>=3.5.0
